Add files via upload
This is a project that required data to be cleaned before creating pivot tables and finally creating a dashboard. This required me to remove duplicate data, fixing the data entered into columns, ensuring that date values were the same, to mention a few(in the data cleaning part of the project).